Move is_safe to controller and fix snake head bug
We only want to consider the space safe if it's adjacent to snake heads that are smaller than us, as that's
a guaranteed kill. If the space we're considering the safete of _currently_ has a snake head, and we move to it,
that snake will also move and the space will then be occupied by a snake body which will kill us.